Refactor runtime loader to avoid merge conflicts
