+++ app/data/__init__.py
@@ -1,6 +1,7 @@
 """Data utilities for the Streamlit prototype."""
 
 from .data_manager import DataManager
+from .runtime_loader import RuntimeContext, load_runtime_context
 from .sample_features import (
     FeatureRecord,
     distinct_values,
@@ -11,6 +12,8 @@ from .sample_features import (
 
 __all__ = [
     "DataManager",
+    "RuntimeContext",
+    "load_runtime_context",
     "FeatureRecord",
     "distinct_values",
     "load_feature_dataframe",
